refactor(dto): WaitingWithRank의 위치를 domain에서 dao/dto로 이동
